Add regression test for snapshot isolation mode.
To test the mode, we use two concurrent transactions generated by 2
pgbench instances. Tx A updates a row of table t1. Tx B retrieves the
row in t1 and INSERT INTO log table. If atomic visibilty is kept,
after Tx A and B finishes rows of log on node 0 and node 1 should be
identical. In fact in the snapshot isolation mode, they are
identical. On the other hand, in the existing native replication mode,
they are not identical.

Also add missing sample configuration file for the snapshot isolation
mode.