Add support for LDAP authentication between client and pgpool.
This commit supports LDAP authentication between application and Pgpool.
LDAP is used only to validate the user name/password pairs. LDAP parameter
(such as ldapserver,port, ldapbinddn, etc..) is specified in a row of
pool_hba.conf. When user connection is matched by ldap authentication rows
in pool_hba.conf, access ldapserver and search matching informatinonfor
username/passord.

pgpool has 'backend_use_password' option which is particular parameter
in auth-options. It is able to reuse authentication password which was
used between application and pgpool to postgres authentication.

If we use this feature, Pgpool-II must be build with --with-ldap support.