Fix for log messages from child processes printing the main process's PID 
The issue was caused by my log collector commit. And the reason was: as part of
the log collector implementation, I added a new global variable myProcPid to hold
to the process PID so that we can avoid calling getpid() function every time
during the construction of the log message.

The change was meant to increase the performance but unfortunately I missed
setting the myProcPid variable after fork() at one of the places.
(After forking the pgpool child process).

So to make sure we can avoid such mistakes in the future, I also did a little bit
of refactoring as part of this fix and created a new function that sets all
process-related global variables (process_application_name, myProcPid and
processType) so that we have a single place to add/modify such stuff.

Apart from that, I have made a few changes in functions for setting the
process_application_name and moved them to pool_globasl.c from pgpool_main.c 