Fix per_node_error_log() so that it respects unread parameter.
per_node_error_log() unconditionally set the "unread" parameter to
true when it should have respected the unread parameter given by user.
It seems all callers to per_node_error_log() set the unread parameter
to true anyway.  The only exception is pool_do_auth. As far as testing
test 03[0-4] regression tests, the change is ok.

In summary this change will not affect to Pgpool-II but bug is bug.