Implement automatic writing function detection by checking volatile property.
If a function included in SELECT/WITH has volatile property by
checking system catalog, regard it a writing function. This feature is
available only when the black_function_list and the
white_function_list are both empty. The default of the
black_function_list and the white_function_list are now empty string
so that this feature is enabled by default.

If the black_function_list or the white_function_list is not empty,
keep on existing behavior (only consult the black_function_list or the
white_function_list, not access system catalog).

Discussion: https://www.pgpool.net/pipermail/pgpool-hackers/2020-July/003730.html
Reviewed-by: Hou, Zhijie