Fix 018.detach_primary test.
The script tried to extract primary node id from the results of show
pool_nodes.  From 4.3 it includes the actual backend status. As a
result the script could accidentally picks up a node id which is
actually a standby (but has primary status as PostgreSQL) from
Pgpool-II's point of view.

This resulted in a shell syntax error here because $primary_node could
have node "0 2" for example.

if [ $primary_node != 0 ];then