New feature "prefer_lower_delay_standby"
Currently on streaming replication mode, when the node selected as the load
balancing node has streaming delay over delay_threshold, pgpool sends the
query to the primary regardless of other standbys.

When this parameter is set to on, if the delayed load balancing node is the
standby, pgpool selects the new load balancing node from the lowest delay standbys.

If all standbys are delayed or the primary is chosen as the load balancing node,
pgpool sends the query to the primary.