Add new fields to pcp_proc_info, show pool_processes, show pool_pools command.
The purpose of this commit is to get more information about the process stats
and how pgpool is working with "client_idle_limit" and "child_max_connections".

Start time:
I add “remaining” that is the remaining time beforl the process is restarted by
client_idle_limit.

Client connection count:
The connection count from clients to this pgpool process. Useful to know how
many times clients can connect to this pgpool process before
child_max_connections reaches.

Backend connection time:
I just rename create_time for clarity.

Client connection time:
The time when the client last connected. If the client uses its own connection
pooling, this helps to know when the connection was actually made from pgpool's
point of view.

Client idle duration:
The elapsed time (in seconds) since the client became in idle state.

Client disconnection time:
This is the last time the client disconnects. The use case is the same as above.

Status:
This is the process status. This shows 4 status: “Wait for connections”, “Idle”,
“Idle in transaction”, “Execute command”.