Attempt to fix buildfarm timeout error.
After inspecting the buildfarm log, I noticed it seems as if terminate
signal was ignored.  My theory is system(3), which is used in
trigger_failover() ignores SIGINT and SIGQUIT.  To deal with this,
check the return value of system(3) to determine whether signal was
sent while executing system(3). If so, call exit_handler after
failover.