Fix pgpool_adm: pcp_pool_status.
If the function is used with PostgreSQL 15, it fails with:

ERROR:  table-function protocol for value-per-call mode was not followed

pcp_pool_status actually returns set of record but it was declared as
"returns record".  This sloppiness had been tolerated until 14 but 15
becomes more strict. To fix this we need to change the return type of
pcp_pool_status from "returns record" to "returns setof record".
So the pgpool_adm extension version is bumped up.